Fix issue with empty clover report xml file
Main script will stop the execution in case the file
located at "$CLOVER_LOCATION" is missing/invalid, though
it is still possible to generate the final html result even
if the clover report is missing.

This change will allow one to run this tool without errors
even if the clover report file is missing. It will create
an empty clover report file in '/tmp' directory.